# Speed up recipe search by item description