--- a/src/player.cpp
+++ b/src/player.cpp
@@ -3736,7 +3736,8 @@ const recipe_subset &player::get_learned_recipes() const
     return *learned_recipes;
 }
 
-recipe_subset player::get_recipes_from_books( const inventory &crafting_inv ) const
+recipe_subset player::get_recipes_from_books( const inventory &crafting_inv,
+        recipe_filter filter ) const
 {
     recipe_subset res;
 
@@ -3745,6 +3746,9 @@ recipe_subset player::get_recipes_from_books( const inventory &crafting_inv ) co
 
         for( std::pair<const recipe *, int> recipe_entry :
              candidate.get_available_recipes( *this ) ) {
+            if( filter && !filter( *recipe_entry.first ) ) {
+                continue;
+            }
             res.include( recipe_entry.first, recipe_entry.second );
         }
     }
@@ -3753,20 +3757,29 @@ recipe_subset player::get_recipes_from_books( const inventory &crafting_inv ) co
 }
 
 recipe_subset player::get_available_recipes( const inventory &crafting_inv,
-        const std::vector<npc *> *helpers ) const
+        const std::vector<npc *> *helpers, recipe_filter filter ) const
 {
-    recipe_subset res( get_learned_recipes() );
+    recipe_subset res;
 
-    res.include( get_recipes_from_books( crafting_inv ) );
+    if( filter ) {
+        res.include_if( get_learned_recipes(), filter );
+    } else {
+        res.include( get_learned_recipes() );
+    }
+
+    res.include( get_recipes_from_books( crafting_inv, filter ) );
 
     if( helpers != nullptr ) {
         for( npc *np : *helpers ) {
             // Directly form the helper's inventory
-            res.include( get_recipes_from_books( np->inv ) );
+            res.include( get_recipes_from_books( np->inv, filter ) );
             // Being told what to do
-            res.include_if( np->get_learned_recipes(), [ this ]( const recipe & r ) {
-                return get_skill_level( r.skill_used ) >= static_cast<int>( r.difficulty *
-                        0.8f ); // Skilled enough to understand
+            res.include_if( np->get_learned_recipes(), [ this, &filter ]( const recipe & r ) {
+                if( filter && !filter( r ) ) {
+                    return false;
+                }
+                // Skilled enough to understand
+                return get_skill_level( r.skill_used ) >= static_cast<int>( r.difficulty * 0.8f );
             } );
         }
     }