#### Purpose of change
I noticed that `d:` search in crafting UI takes enormous amount of time, particularly for characters that know all recipes.
In VS release build, it takes 27 seconds. Changes in this PR cut that number down to 4 seconds.

#### Describe the solution
When profiling, I noticed that approximately 70% of time is spent on formatting `You could use it to craft` item info section. Out of those 70%, 50% is spent on creating new `recipe_subset` from list of known recipes, and remaining 20% is spent on destroying the subset.

I've managed to bring the time down to 5 seconds by filtering out recipes that don't have the item as component before adding them to the subset.

Additional 1 second was squeezed out by getting rid of vector allocation in `Item_factory::find`, and iterating over original vectors instead.

#### Describe alternatives you've considered
Out of remaining time, approximately third is spent on finding holster iuse within all items, and another third on calling `res.include_if( get_learned_recipes(), filter )`, so there is definitely room for improvement.

#### Testing
In curses release build, searching for `d:longp` across all recipes is completed in 4 seconds (down from 11).