@olanti-p 2021-09-21
It may be, though it wouldn't surprise me if they already did something similar.