@Coolthulhu 2021-09-22
> Additional 1 second was squeezed out by getting rid of vector allocation in Item_factory::find, and iterating over original vectors instead.

So `Item_factory::all()` recreates the result every single time instead of caching it and only rebuilding it when dirty. I guess it worked so far, but it must have been pain for users who stack uncountable infinity of mods to their game.

> approximately third is spent on finding holster iuse within all items

It goes over all item types that exist, just to find ones with holster action. And doesn't cache anything at all.
`Item_factory::register_cached_uses` does something like this except for repair actions.